+++ src/AudioInterface.h
@@ -16,7 +16,7 @@ public:
 
     static const int BUFFER_LEN = 4096;
 
-    virtual void AudioDataReady(const std::array<uint8_t, BUFFER_LEN> &data) = 0;
+    virtual void AudioDataReady(const std::array<int16_t, BUFFER_LEN> &data) = 0;
     virtual int GetAudioSampleRate() = 0;
     virtual bool GetAudioEnabled() = 0;
     virtual Channels GetEnabledAudioChannels() = 0;
